Move api logic into separate file. add complete_item route and tick button
